[codex] index slice coverage and stream large blobs
slice_included_paths table replaces per-path full-table coverage scans
with one batched ancestor-prefix lookup per patchset. UploadBlobStream/
ReadBlobStream add hash-verified streaming transfer; CLI and hydration
use streaming above 4MiB. Operator fixed the hydration assertion to use
the canonical workspace layout and ran the real-Postgres e2e gates.

+++ service/metrics.go
@@ -13,7 +13,7 @@ var (
 	)
 )
 
-func recordBlobUpload(size int) {
+func recordBlobUpload(size int64) {
 	blobUploadsTotal.Inc(nil)
 	blobUploadBytesTotal.Add(float64(size), nil)
 }
